procedures complete READ DESCRIPTION!!!!
found a full explanation of the EBNF for everything in the lecture 7
notes, check those out to go over your own code and then we will have to
add more stuff to our parser to flesh it all out. So we need to discuss
the further distribution of labor in the groupme.